refactor: fix changes for refactoring engine (#2)
