+++ .eslintrc.json
@@ -1,4 +1,5 @@
 {
+  "extends": "next/core-web-vitals",
   "parserOptions": {
     "sourceType": "module"
   },
