--- src/pages/engine/shared.ts
@@ -1,1 +1,0 @@
-export const STORAGE_CACHE_DIR = './cache';
